Feature: implement SObject interfaces.
This change implements the SObject interfaces. An SObject is defined by
Salesforce API to wrap data and operations of an object, or record, on
Salesforce database.

An SObject must be associated with an instance of Client to talk to the
Salesforce server and exchange SObject data. An SObject can be created
in 3 ways:

- client.SObject(typeName string) returns an SObject with a provided
  typeName.
- client.Query(q string) returns a QueryResult, which may contain a list
  of SObjects in the records field;
- sobject.SObjectField(typeName, key string) returns an SObject instance
  with the provided typeName and the external ID indexed by "key" within
  the invoking sobject. This is convenient to look up for the linked
  object or parent object.

The following operations are defined but to-be-implemented:

- sobject.Get() shall invoke the REST API to retrieve detailed data
  fields of an SObject. The "Id" field must exist for this operation to
  work;
- sobject.Describe() shall invoke the REST API to retrieve metadata of
  the SObject type. The "Type" field in "attributes" must exist.

The following operations are yet-to-be-defined:

- sobject.Create()
- sobject.Update()
- sobject.Delete()