Support linked SObjects returned by Query.
In the query response, when there're linked objects, the object is
returned as nested JSON dictionary, without Id key provided. In this
case, the code should check for the availability of "attributes" in a
nested JSON object, and decode valid ID from attributes.url, and
generate a proper SObject with the SObject.SObjectField() method.